moved Consumer and Acceptor, as well as consumers and acceptors, to their own files
